requirements.txt
discord==2.3.1
discord.py==2.3.1
python-telegram-bot==12.5